Add /r/nfl
Angry_Caveman_Lawyer:
It depends on the bot. Tweet poster for example is allowed and encouraged. A wiki bot would be banned.
skepticismissurvival:
Basically, we base whether or not we allow bots on whether or not the bots would actually be useful and offer a significant benefit to subreddit.
----------------------
Thanks!